Develop use-case List Tarif dan Brg

+++ Bilreg.Application/ChargeContext/TarifFeature/IStokRepo.cs
@@ -8,4 +8,4 @@ public interface IStokRepo :
 {
 }
 
-public record StokView(string BrgId, string BrgName, int Qty, string Satuan);
+public record StokView(string BrgId, string BrgName, decimal Qty, string Satuan);
